requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
PyAudio==0.2.13
PyYAML==6.0
requests==2.28.2
SpeechRecognition==3.9.0
urllib3==1.26.14
